refactor(rpc): extract prepare_env logic
